release - 0.0.22 (#44)
* fix : 일간 정보 확인 API에 워커와 기버에 userType 필드값이 필요없어서 삭제 (#2)

* add cicd workflow (#1)

* add cd workflow

* cd: add missed login ghcr

* cd: add dockerfile path

* cd: set context

* add docker-compose and test

* revert swagger version string

* 일감 정보 확인 API userId -> applicationId로 변경 및 personnel 필드값 삭제 (#5)

* fix : 일간 정보 확인 API 응답값에 personnel 기획상 필요없다는 하셔서 삭제

* fix : 일간 정보 확인 API userId가 아닌 applicationId로 받도록 변경

* rename and cleanup regex for nickname (#6)

* update update user (#7)

* refactor tag service, controller for ResponseEntity (#8)

* add id for getTagRank (#10)

* sw-753 groupstatus update (#9)

* sw-753 groupStatus update

* sw-753 groupStatus update

* status update

* status update

* 안쓰는거 삭제 및 수정

* groupstatus -> groupStatus 통일

* Update src/main/java/com/sharework/response/model/job/APIProceedingList.java


* fix typo error

* fix typo groupstatus

---------


* update user update (#13)

* Add checklist and tags for getPreviousJobs (#12)

* rename to JobPreviousPayload for swagger

* add checklist for getPreviousJobs

* remove jobBenefits for getPreviousJobs

* add tags for getPreviousJobs

* rename to jobChecklistContents

* fix build error

* update groupStatus (#16)

* fix response model for swagger (#15)

* fix wrong JobHiredInfoPayload (#17)

* add personnel for getPreviousJobs (#18)

* review update (#19)

* review update

* Update src/main/java/com/sharework/dao/ReviewDao.java


* Update src/main/java/com/sharework/dao/BaseReviewDao.java


---------


* GET application isReviewed add and receiptwoker delete and Giver rece… (#20)

* GET application isReviewed add and receiptwoker delete and Giver receipt isReviewed add.

* code clean

* Fix updateRejected to single and 3 hours (#21)

* fix updateRejected to single and 3 hours

* fix meta message

* fix wrong reject flow (#22)

* refactor tag and user for ResponseEntity (#23)

* feat: add slack alerts for dev (#25)

* Refactor registration for ResponseEntity (#24)

* refactor registration for ResponseEntity

* rename and comments

* refactor exception handler in RegistrationController

fix try-catch to ControllerAdvice
replace ErrorResponse with NotFoundException

* update sharework-profiles (#26)

* refactor checklist, sms for ResponseEntity (#27)

* nickname check msg update (#28)

* nickname check msg update

* success msg update

* fix swagger response for application and baseReview (#29)

* Add checkNickname test and JaCoCo plugin (#30)

* add checkNickname test

* add jacoco for coverage

* open -> open,closed로  변경 (#31)

* fix date field for RegisterJob (#32)

* Refactor ResponseEntity for Application and BaseReview (#33)

* refactor ResponseEntity for Application

* refactor ResponseEntity for BaseReview

* fix swagger response for JobController (#35)

* Add alarm simple feature (#34)

* add alarm simple feature

* update sharework-profiles

* Fix fcm path to ClassPathResource (#36)

* fix fcm path to ClassPathResource

* update sharework-profiles

* Fix error not found class path in local (#37)

* fix error not found class path in local

* fix field injection to parameter for Autowired

* add FirebaseConfigTest

* Add feature flyway and init script (#38)

* add feature flyway and init script

* update sharework-profiles

* update sharework-profiles (#39)

* rename db column jwt in User (#40)

* Add prod CICD (#41)

* fix cicd trigger for dev

* add prod cicd and fix swagger desc.

* update sharework-profiles

* fix cicd and add docker-compose (#42)

* release - 0.0.22 (#43)

---------

+++ src/main/java/com/sharework/service/UserService.java
@@ -9,6 +9,7 @@ import com.amazonaws.services.s3.model.PutObjectRequest;
 import com.sharework.common.ApplicationTypeEnum;
 import com.sharework.common.JobTypeEnum;
 import com.sharework.dao.*;
+import com.sharework.global.NotFoundException;
 import com.sharework.manager.CreateJwt;
 import com.sharework.manager.HashidsManager;
 import com.sharework.manager.JwtManager;
@@ -18,16 +19,15 @@ import com.sharework.request.model.APIUpdateUser;
 import com.sharework.request.model.LoginObj;
 import com.sharework.request.model.SignInRequestPw;
 import com.sharework.request.model.SignupRequestPw;
-import com.sharework.response.model.*;
+import com.sharework.response.model.SuccessResponse;
+import com.sharework.response.model.VerifiedPayload;
+import com.sharework.response.model.VerifiedResponse;
 import com.sharework.response.model.job.JobTagList;
 import com.sharework.response.model.meta.BasicMeta;
-import com.sharework.response.model.tag.APIGetTagRank;
 import com.sharework.response.model.tag.JobTagRank;
 import com.sharework.response.model.tag.TagRank;
-import com.sharework.response.model.user.APIGetUser;
 import com.sharework.response.model.user.Giver;
 import com.sharework.response.model.user.Profile;
-import javassist.NotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
@@ -39,6 +39,7 @@ import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.server.ResponseStatusException;
 
+import javax.transaction.Transactional;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
@@ -123,44 +124,20 @@ public class UserService {
 //        return response;
 //    }
 
-    public ResponseEntity signUpPw(SignupRequestPw request, BindingResult bindingResult) {
-
-        ResponseEntity response = null;
-        ErrorResponse error = null;
-        String errorMsg = null;
-        BasicMeta meta;
-
-        // vaild가 틀렸을 경우
+    public SuccessResponse signupPw(SignupRequestPw request, BindingResult bindingResult) {// vaild가 틀렸을 경우
         if (bindingResult.hasErrors()) {
-            errorMsg = "정확한 정보를 기입해주세요.";
-            meta = new BasicMeta(false, errorMsg);
-            error = new ErrorResponse(meta);
-            response = new ResponseEntity<>(error, HttpStatus.OK);
-            return response;
+            return new SuccessResponse(new BasicMeta(false, "정확한 정보를 기입해주세요."));
         } else if (userDao.existsByEmailAndDeleteYn(request.getEmail(), "N")) {
-            errorMsg = "이메일이 중복됩니다.";
-            meta = new BasicMeta(false, errorMsg);
-            error = new ErrorResponse(meta);
-            response = new ResponseEntity<>(error, HttpStatus.OK);
-            return response;
+            return new SuccessResponse(new BasicMeta(false, "이메일이 중복됩니다."));
         } else if (userDao.existsByPhoneNumberAndDeleteYn(request.getPhoneNumber(), "N")) {
-            errorMsg = "번호가 중복됩니다.";
-            meta = new BasicMeta(false, errorMsg);
-            error = new ErrorResponse(meta);
-            response = new ResponseEntity<>(error, HttpStatus.OK);
-            return response;
+            return new SuccessResponse(new BasicMeta(false, "번호가 중복됩니다."));
         } else if (userDao.existsByNameAndDeleteYn(request.getName(), "N")) {
-            errorMsg = "닉네임이 중복됩니다.";
-            meta = new BasicMeta(false, errorMsg);
-            error = new ErrorResponse(meta);
-            response = new ResponseEntity<>(error, HttpStatus.OK);
-            return response;
+            return new SuccessResponse(new BasicMeta(false, "이미 사용중인 닉네임 입니다. 다른 닉네임을 입력해 주세요."));
         }
 
         ResidentNumberJsonb residentNumber = null;
 
         if (request.getResidentNumberFront() != null && request.getResidentNumberRear() != null) {
-            System.out.println(123);
             residentNumber = new ResidentNumberJsonb(request.getResidentNumberFront(),
                     request.getResidentNumberRear());
         }
@@ -169,21 +146,10 @@ public class UserService {
         userDao.save(
                 User.builder().email(request.getEmail()).name(request.getName()).phoneNumber(request.getPhoneNumber())
                         .residentNumber(residentNumber).userType("worker").password(request.getPassword()).refreshToken(refreshToken).build());
-
-        meta = new BasicMeta(true, "회원가입이 완료되었습니다.");
-        SuccessResponse successResponse = new SuccessResponse(meta);
-        response = new ResponseEntity<>(successResponse, HttpStatus.OK);
-        return response;
+        return new SuccessResponse(new BasicMeta(true, "회원가입이 완료되었습니다."));
     }
 
-    public ResponseEntity login(SignInRequestPw request) {
-
-        ResponseEntity response = null;
-        ErrorResponse error = null;
-        String errorMsg = null;
-        BasicMeta meta;
-
-
+    public VerifiedResponse login(SignInRequestPw request) {
         Optional<User> user = userDao.findUserByPhoneNumberAndPasswordAndDeleteYn(request.getPhoneNumber(), request.getPassword(), "N");
         LoginObj loginObj = new LoginObj();
 
@@ -202,46 +168,24 @@ public class UserService {
 
         if (loginObj.getFlag()) {
             VerifiedPayload verifiedPayload = new VerifiedPayload(loginObj.getAccessToken(), loginObj.getRefreshToken(), loginObj.getUserType());
-            meta = new BasicMeta(true, "로그인 성공");
-            VerifiedResponse verifiedResponse = new VerifiedResponse(verifiedPayload, meta);
-            response = new ResponseEntity<>(verifiedResponse, HttpStatus.OK);
+            BasicMeta meta = new BasicMeta(true, "로그인 성공");
+
+            return new VerifiedResponse(verifiedPayload, meta);
         } else {
-            errorMsg = "로그인 정보가 일치하지 않습니다.";
-            meta = new BasicMeta(false, errorMsg);
-            error = new ErrorResponse(meta);
-            response = new ResponseEntity<>(error, HttpStatus.OK);
+            throw new NotFoundException("로그인 정보가 일치하지 않습니다.");
         }
-        return response;
     }
 
-    public ResponseEntity checkNickname(String nickName) {
-
-        ResponseEntity response = null;
-        ErrorResponse error = null;
-        String errorMsg = null;
-        BasicMeta meta;
-
-
-        if (!Pattern.matches("^[가-힣a-zA-Z0-9]{1,6}$", nickName)) {
-            errorMsg = "사용 불가능한 닉네임입니다.";
-            meta = new BasicMeta(false, errorMsg);
-            error = new ErrorResponse(meta);
-            response = new ResponseEntity<>(error, HttpStatus.OK);
-            return response;
+    public SuccessResponse checkNickname(String nickname) {
+        if (!Pattern.matches("^[가-힣a-zA-Z0-9]{1,6}$", nickname)) {
+            return new SuccessResponse(new BasicMeta(false, "사용할 수 없는 닉네임입니다. 한글, 영문, 숫자로 구성된 6자이내 닉네임을 입력해 주세요"));
         }
 
-        if (userDao.existsByNameAndDeleteYn(nickName, "N")) {
-            errorMsg = "닉네임이 존재합니다.";
-            meta = new BasicMeta(false, errorMsg);
-            error = new ErrorResponse(meta);
-            response = new ResponseEntity<>(error, HttpStatus.OK);
-            return response;
+        if (userDao.existsByNameAndDeleteYn(nickname, "N")) {
+            return new SuccessResponse(new BasicMeta(false, "이미 사용중인 닉네임 입니다. 다른 닉네임을 입력해 주세요."));
         }
 
-        meta = new BasicMeta(true, "사용 가능한 닉네임입니다.");
-        SuccessResponse successResponse = new SuccessResponse(meta);
-        response = new ResponseEntity<>(successResponse, HttpStatus.OK);
-        return response;
+        return new SuccessResponse(new BasicMeta(true, "사용 가능한 닉네임입니다."));
     }
 
     @ExceptionHandler(NotFoundException.class)
@@ -249,10 +193,7 @@ public class UserService {
         return ResponseEntity.status(HttpStatus.NOT_FOUND).body(ex.getMessage());
     }
 
-    public ResponseEntity getUser(String accessToken) {
-        ResponseEntity response = null;
-        Response error = null;
-
+    public Profile getUser(String accessToken) {
         long userId = identification.getHeadertoken(accessToken);
 
         User user = userDao.findByIdAndDeleteYn(userId, "N").orElseThrow(() -> new ResponseStatusException(HttpStatus.OK));
@@ -265,36 +206,23 @@ public class UserService {
         userRate.ifPresent(userRating -> {
             profile.setRate(userRating.getRate());
         });
-        BasicMeta meta = new BasicMeta(true, "");
-        APIGetUser apiGetUser = new APIGetUser(new APIGetUser.Payload(profile), meta);
-        response = new ResponseEntity<>(apiGetUser, HttpStatus.OK);
-        return response;
+
+        return profile;
     }
 
-    public ResponseEntity updateUser(String accessToken, APIUpdateUser request) {
-
-        ResponseEntity response = null;
-        Response error = null;
-
+    @Transactional
+    public void updateUser(String accessToken, APIUpdateUser request) {
         long userId = identification.getHeadertoken(accessToken);
 
-        Optional<User> user = userDao.findByIdAndDeleteYn(userId, "N");
+        User user = userDao.findByIdAndDeleteYn(userId, "N").orElseThrow();
 
-        user.get().setName(request.getName());
-        user.get().setComment(request.getComment());
-        userDao.save(user.get());
-
-        BasicMeta meta = new BasicMeta(true, "정보가 수정되었습니다.");
-        SuccessResponse successResponse = new SuccessResponse(meta);
-        response = new ResponseEntity<>(successResponse, HttpStatus.OK);
-        return response;
+        if(request.getName() != null)
+           user.setName(request.getName());
+        if(request.getComment() != null)
+           user.setComment(request.getComment());
     }
 
-    public ResponseEntity getTagRank(String accessToken) {
-
-        ResponseEntity response = null;
-        Response error = null;
-
+    public List<TagRank> getTagRank(String accessToken) {
         long userId = identification.getHeadertoken(accessToken);
 
         User user = userDao.findByIdAndDeleteYn(userId, "N").orElseThrow();
@@ -311,10 +239,10 @@ public class UserService {
                 jobIdList.add(job.getId());
             });
 
-            jobTagRankList = jobTagDao.findByJobIdCountContents(jobIdList);
+            jobTagRankList = jobTagDao.findByJobIdCountContentsId(jobIdList);
 
             jobTagRankList.forEach(jobTagRank -> {
-                tagRankList.add(new TagRank(new JobTagList(jobTagRank.getContents()), jobTagRank.getCount()));
+                tagRankList.add(new TagRank(new JobTagList(jobTagRank.getId(), jobTagRank.getContents()), jobTagRank.getCount()));
             });
         } else {
             List<Application> applicationList = applicationDao.getByUserIdAndStatus(userId, ApplicationTypeEnum.COMPLETED.name());
@@ -323,24 +251,18 @@ public class UserService {
                 jobIdList.add(application.getJobId());
             });
 
-            jobTagRankList = jobTagDao.findByJobIdCountContents(jobIdList);
+            jobTagRankList = jobTagDao.findByJobIdCountContentsId(jobIdList);
 
             jobTagRankList.forEach(jobTagRank -> {
                 int hour = applicationDao.countByUserIdAndTagContents(userId, jobTagRank.getContents());
-                tagRankList.add(new TagRank(new JobTagList(jobTagRank.getContents()), jobTagRank.getCount(), hour));
+                tagRankList.add(new TagRank(new JobTagList(jobTagRank.getId(), jobTagRank.getContents()), jobTagRank.getCount(), hour));
             });
         }
 
-        BasicMeta meta = new BasicMeta(true, "");
-        APIGetTagRank apiGetTagRank = new APIGetTagRank(new APIGetTagRank.Payload(tagRankList), meta);
-        response = new ResponseEntity<>(apiGetTagRank, HttpStatus.OK);
-        return response;
+        return tagRankList;
     }
 
-    public ResponseEntity insertImg(String accessToken, MultipartFile multipartFile) {
-        ResponseEntity response = null;
-        Response error = null;
-
+    public boolean insertImg(String accessToken, MultipartFile multipartFile) {
         long userId = identification.getHeadertoken(accessToken);
         userDao.findByIdAndDeleteYn(userId, "N").ifPresent(selectUser -> {
             if (selectUser.getProfileImg() != null)
@@ -379,52 +301,35 @@ public class UserService {
         }
 
         //user 프로필 url 저장.
-        final BasicMeta meta = new BasicMeta();
-        userDao.findByIdAndDeleteYn(userId, "N").ifPresentOrElse(user -> {
+        Optional<User> userOptional = userDao.findByIdAndDeleteYn(userId, "N");
+        if (userOptional.isPresent()) {
+            User user = userOptional.get();
             String imagePath = amazonS3.getUrl(bucket, title).toString();
             user.setProfileImg(imagePath);
             user.setImgTitle(title);
             userDao.save(user);
-            meta.setStatus(true);
-            meta.setMessage("이미지가 성공적으로 저장되었습니다.");
-        }, () -> {
-            meta.setStatus(false);
-            meta.setMessage("유저 정보가 존재하지 않습니다.");
-        });
-
-        SuccessResponse result = new SuccessResponse(meta);
-        response = new ResponseEntity<>(result, HttpStatus.OK);
-        return response;
+            return true;
+        } else {
+            return false;
+        }
     }
 
-    public ResponseEntity withDrawal(String accessToken, String refreshToken) {
-        ResponseEntity response = null;
-        Response error = null;
-
+    public SuccessResponse withdrawal(String accessToken, String refreshToken) {
         long userId = identification.getHeadertoken(accessToken);
         User user = userDao.findByIdAndDeleteYn(userId, "N").orElseThrow();
-        String deleteYn = user.getDeleteYn();
+        String deleteYn = user.getDeleteYn(); // FIXME: already an exception above.
 
         if (deleteYn.equals("Y")) {
-            String errorMsg = "삭제된 회원입니다.";
-            error = new Response(new BasicMeta(false, errorMsg));
-            response = new ResponseEntity<>(error, HttpStatus.OK);
-            return response;
+            return new SuccessResponse(new BasicMeta(false, "삭제된 회원입니다."));
         }
 
-        BasicMeta meta = new BasicMeta();
-
-        boolean flag = false;
         // 일감 오픈 및 진행중이라면 탈퇴불가
         ArrayList<String> jobStatusList = new ArrayList<>();
         jobStatusList.add(JobTypeEnum.OPEN.name());
         jobStatusList.add(JobTypeEnum.CLOSED.name());
         jobStatusList.add(JobTypeEnum.STARTED.name());
         if (jobDao.findByUserIdAndStatusIn(userId, jobStatusList).size() > 0) {
-            String errorMsg = "진행중인 일감 삭제 후 탈퇴진행해 주시기 바랍니다.";
-            error = new Response(new BasicMeta(false, errorMsg));
-            response = new ResponseEntity<>(error, HttpStatus.OK);
-            return response;
+            return new SuccessResponse(new BasicMeta(false, "진행중인 일감 삭제 후 탈퇴진행해 주시기 바랍니다."));
         }
 
         // hired_request,,hired_approved 일 마감하라는 에러 메시지 출력.
@@ -432,10 +337,7 @@ public class UserService {
         applicationFailedStatusList.add(ApplicationTypeEnum.HIRED_REQUEST.name());
         applicationFailedStatusList.add(ApplicationTypeEnum.HIRED_APPROVED.name());
         if (applicationDao.countByUserIdAndStatusIn(userId, applicationFailedStatusList) > 0) {
-            String errorMsg = "진행중인 일 처리 후 탈퇴진행해 주시기 바랍니다.";
-            error = new Response(new BasicMeta(false, errorMsg));
-            response = new ResponseEntity<>(error, HttpStatus.OK);
-            return response;
+            return new SuccessResponse(new BasicMeta(false, "진행중인 일 처리 후 탈퇴진행해 주시기 바랍니다."));
         }
 
         ArrayList<String> applicationSuccessStatusList = new ArrayList<>();
@@ -449,66 +351,43 @@ public class UserService {
             userDao.save(users);
         });
 
-        meta = new BasicMeta(true, "회원탈퇴가 성공하였습니다.");
-        SuccessResponse result = new SuccessResponse(meta);
-        response = new ResponseEntity<>(result, HttpStatus.OK);
-        return response;
+        return new SuccessResponse(new BasicMeta(true, "회원탈퇴가 성공하였습니다."));
     }
 
-    public ResponseEntity getImg(String accessToken) {
-        ResponseEntity response = null;
-        Response error = null;
-
+    public String getImg(String accessToken) {
         long userId = identification.getHeadertoken(accessToken);
 
-        final ImgPayload imgPayload = new ImgPayload();
-        final BasicMeta meta = new BasicMeta();
-        userDao.findByIdAndDeleteYn(userId, "N").ifPresentOrElse(profileUser -> {
-            imgPayload.setProfileImg(profileUser.getProfileImg());
-            meta.setStatus(true);
-            meta.setMessage("");
-        }, () -> {
-            meta.setStatus(false);
-            meta.setMessage("유저 정보가 존재하지 않습니다.");
-        });
-
-        ImgResponse imgResponse = new ImgResponse(meta, imgPayload);
-        response = new ResponseEntity<>(imgResponse, HttpStatus.OK);
-        return response;
+        Optional<User> userOptional = userDao.findByIdAndDeleteYn(userId, "N");
+        if (userOptional.isPresent()) {
+            User user = userOptional.get();
+            return user.getProfileImg();
+        } else {
+            return "";
+        }
     }
 
-    public ResponseEntity deleteImg(String accessToken) {
-        ResponseEntity response = null;
-        Response error = null;
-
+    public boolean deleteImg(String accessToken) {
         long userId = identification.getHeadertoken(accessToken);
 
-        final BasicMeta meta = new BasicMeta();
-        userDao.findByIdAndDeleteYn(userId, "N").ifPresentOrElse(profileUser -> {
-            String fileName = profileUser.getImgTitle();
+        Optional<User> userOptional = userDao.findByIdAndDeleteYn(userId, "N");
+        if (userOptional.isPresent()) {
+            User user = userOptional.get();
+            String fileName = user.getImgTitle();
 
             //이미지가 존재하지 않을 경우.
             if (fileName == null) {
-                meta.setStatus(false);
-                meta.setMessage("이미지가 존재하지 않습니다.");
-                return;
+                return false;
             }
 
             amazonS3.deleteObject(bucket, fileName);
 
             //user profleImg null로 변경
-            profileUser.setProfileImg(null);
-            profileUser.setImgTitle(null);
-            userDao.save(profileUser);
-            meta.setStatus(true);
-            meta.setMessage("성공적으로 삭제하였습니다.");
-        }, () -> {
-            meta.setStatus(false);
-            meta.setMessage("유저 정보가 존재하지 않습니다.");
-        });
-
-        SuccessResponse successResponse = new SuccessResponse(meta);
-        response = new ResponseEntity<>(successResponse, HttpStatus.OK);
-        return response;
+            user.setProfileImg(null);
+            user.setImgTitle(null);
+            userDao.save(user);
+            return true;
+        } else {
+            return false;
+        }
     }
 }
